+++ packages/@rescript/runtime/lib/js/Stdlib_Result.js
@@ -341,6 +341,60 @@ function all6(param) {
   }
 }
 
+async function mapOkAsync(res, f) {
+  let value = await res;
+  if (value.TAG === "Ok") {
+    return {
+      TAG: "Ok",
+      _0: f(value._0)
+    };
+  } else {
+    return {
+      TAG: "Error",
+      _0: value._0
+    };
+  }
+}
+
+async function mapErrorAsync(res, f) {
+  let value = await res;
+  if (value.TAG === "Ok") {
+    return {
+      TAG: "Ok",
+      _0: value._0
+    };
+  } else {
+    return {
+      TAG: "Error",
+      _0: f(value._0)
+    };
+  }
+}
+
+async function flatMapOkAsync(res, f) {
+  let value = await res;
+  if (value.TAG === "Ok") {
+    return await f(value._0);
+  } else {
+    return {
+      TAG: "Error",
+      _0: value._0
+    };
+  }
+}
+
+async function flatMapErrorAsync(res, f) {
+  let value = await res;
+  if (value.TAG === "Ok") {
+    return {
+      TAG: "Ok",
+      _0: value._0
+    };
+  } else {
+    return await f(value._0);
+  }
+}
+
 let getExn = getOrThrow;
 
 let mapWithDefault = mapOr;
@@ -367,4 +421,8 @@ exports.all3 = all3;
 exports.all4 = all4;
 exports.all5 = all5;
 exports.all6 = all6;
+exports.mapOkAsync = mapOkAsync;
+exports.mapErrorAsync = mapErrorAsync;
+exports.flatMapOkAsync = flatMapOkAsync;
+exports.flatMapErrorAsync = flatMapErrorAsync;
 /* No side effect */